backend/pyproject.toml
[project]
name = "flight-assistance-backend"
version = "0.1.0"
description = "Backend for Flight Assistance Application"
requires-python = ">=3.9"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn>=0.34.0,<0.35.0",
  "python-dotenv==1.1.0",
  "langchain==0.3.25",
  "langchain-community==0.3.24",
  "langchain-core==0.3.63",
  "langchain-openai==0.3.18",
  "openai==1.82.1",
  "httpx==0.28.1",
  "requests==2.32.3",
  "pydantic>=2.0.0,<3.0.0",
  "PyYAML==6.0.2",
  "regex==2024.11.6",
  "SQLAlchemy==2.0.41",
  "tiktoken==0.9.0",
  "typing_extensions==4.13.2",
  "urllib3==2.4.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["services"]